% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/startle.r
\name{basicStartlePlot}
\alias{basicStartlePlot}
\title{Basic Startle Plot}
\usage{
basicStartlePlot(df, filter_groups, n_col)
}
\arguments{
\item{df}{A dataframe returned by loadStartleData function.}

\item{filter_groups}{An optional argument, a vector of strings used for filtering the data and displaying only wished groups on plots.}

\item{n_col}{An optional argument, defines number of graph columns.}
}
\value{
Returns a time series plot.
}
\description{
Returns a simple time series plot.
}
\examples{
# Load example data
df <- ratpasta_demo

# running basicStartlePlot
basicStartlePlot(df)

}
