% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gripasta.r
\name{errors}
\alias{errors}
\title{Artefact identifier}
\usage{
errors(df, usePlotly)
}
\arguments{
\item{df}{data}

\item{usePlotly}{TRUE or FALSE argument. If set to TRUE, plotly will be used to render plot for each animal.}
}
\value{
Returns a processed griPASTA data
}
\description{
A function that behaves similar to a loop - reruns itself until user reports that no artifacts are present. If artifact is present, user is promted to enter starting and ending x coordinate of the artefact, and artefact is deleted.
}
