% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gripasta.r
\name{findPeaks}
\alias{findPeaks}
\title{Find peaks}
\usage{
findPeaks(df, n, range)
}
\arguments{
\item{df}{Data from griPASTA}

\item{n}{Number of trials}

\item{range}{Estimated length of each trial in seconds.}
}
\value{
returns values of peaks
}
\description{
Finds
}
