% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gripasta.r
\name{griPASTA}
\alias{griPASTA}
\title{griPASTA main function}
\usage{
griPASTA(data, usePlotly = TRUE)
}
\arguments{
\item{data}{Raw griPASTA data loaded with function loadgriPASTA or manually.}

\item{usePlotly}{TRUE or FALSE argument. If set to TRUE, plotly will be used to render plot for each animal.}
}
\value{
Returns measurements for each animal
}
\description{
A function that takes in raw griPASTA data and behaves similar to a loop - continuously promts user exclude artifacts until no artifacts are present. If artifact is present, user is promted to enter starting and ending x coordinate of the artefact, and artefact is deleted.
}
