% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/startle.r
\name{latencyPlot}
\alias{latencyPlot}
\title{Latency plot}
\usage{
latencyPlot(df, addhead)
}
\arguments{
\item{df}{A dataframe returned by loadStartleData function.}

\item{addhead}{Optional argument. If this argument is used in loadStartleData, set addhead to the value}
}
\value{
List with two plots
}
\description{
Returns two plots displaying latency in a list. Latency refers to the reaction of the test subject to the startling sound. It is time period from the startling sound to the time stamp of maximal values recorded.
}
\examples{
# Load data provided with example
df <- ratpasta_demo

# running examples
l <- latencyPlot(df, addhead = 0.1)
}
