% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gripasta.r
\name{loadgriPASTA}
\alias{loadgriPASTA}
\title{Load griPASTA Data}
\usage{
loadgriPASTA()
}
\value{
A dataframe with all startle data
}
\description{
This function is used for loading and processing all data related to startle experiments. All data generated with PASTA, Platform for Acoustic STArle experiments, should be placed in a single folder. Navigate to this folder, set it as working directory, and call this function. All data will be loaded, processed and assigned to a chosen variable that can be passed on other functions.
}
