% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/startle.r
\docType{data}
\name{ratpasta_demo}
\alias{ratpasta_demo}
\title{Sample}
\format{
dataframe
}
\usage{
ratpasta_demo
}
\description{
A sample data for example. A small part of data used in vignette.
}
\keyword{datasets}
