% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/startle.r
\name{loadStartleData}
\alias{loadStartleData}
\title{Load Startle Data}
\usage{
loadStartleData(
  local_import = TRUE,
  group_names,
  addhead,
  addtail,
  metadata,
  correction = TRUE,
  synchronise = FALSE
)
}
\arguments{
\item{local_import}{An argument with default value TRUE. If set to TRUE it will load all .pasta files from the working directory and merge them apropriately. If set to FALSE, a user must specify group_names.}

\item{group_names}{An argument used only if local_import = FALSE. A vector of strings with names of .pasta files pertaining to startle data.}

\item{addhead}{Optional argument. Adds a number of seconds to the duration of the impulse. Acounts for the lag of the animal.}

\item{addtail}{Optional argument. Adds a number of seconds to the duration of the impulse. Acounts for the lag of the animal.}

\item{metadata}{Optional argument. For a custom made pulse protocol.}

\item{correction}{Argument with default value TRUE. If set to TRUE, values will be corrected for the animal mass.}

\item{synchronise}{Argument with default value FALSE. If set to TRUE, timestamps will be corrected based on the synchroniseTime.csv in the working directory.}
}
\value{
A dataframe with all startle data
}
\description{
This function is used for loading and processing all data related to startle experiments. All data generated with PASTA, Platform for Acoustic STArle experiments, should be placed in a single folder. Navigate to this folder, set it as working directory, and call this function. All data will be loaded, processed and assigned to a chosen variable that can be passed on other functions.
}
\examples{
\dontrun{
df <- loadStartleData(df, addhead = 0.2, addtail = 0.2)
# OR
df <- loadStartleData(local_import=FALSE, group_names=c("ctr 1", "ctr 2"), addhead=0.1, addtail=1)
}
}
